What about an on-demand GitHub Action?
What about an on-demand GitHub Action that would create a pull request on another GitHub repo?

`itinerant_codespell` is a simple on-demand GitHub Action that allows a user to enter the _user_name/repo_name_ of any repo on GitHub.  The Action clones that repo and then runs `codespell` on it.  This is a trivial example but any tool could be used in place of `codespell` and the idea could be extended to create a PR containing the changes made back to the remote repository.

A similar on-demand `itinerant_remove_print_statements.yml` could enable users to get all the advantages of this tool without installing it locally.

Thoughts?